requirements.txt
# requirements.txt
gradio
python-dotenv
pandas
numpy
binance-connector
python-binance
requests
plotly
pandas-ta
